feat: React 19 support (2)
