+++ packages/fastify-react/package.json
@@ -42,6 +42,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-router": "^7.5.0",
+    "react-router-dom": "^7.5.0",
     "unihead": "latest",
     "valtio": "latest",
     "youch": "^3.3.4"
