+++ packages/fastify-react/virtual/root.jsx
@@ -3,7 +3,6 @@ import { Route, Routes } from 'react-router-dom'
 import { AppRoute, Router } from '$app/core.jsx'
 
 export default function Root({ url, routes, head, ctxHydration, routeMap }) {
-  console.log('routes', routes)
   return (
     <Suspense>
       <Router location={url}>
