+++ starters/react-base/postcss.config.js
@@ -1,0 +1,12 @@
+export default {
+  plugins: {
+    '@tailwindcss/postcss': {},
+    'postcss-preset-env': {
+      stage: 1,
+      features: {
+        // Let Tailwind handle it
+        'nesting-rules': false
+      }
+    },
+  }
+}
